# Improve pgn parser performance