Follow @johnynek suggestions in #239. The the performance is improved dramatically!
Here is a typical benchmark result in my machine:
```
55 games in 22 ms
55 games in 31 ms
55 games in 25 ms
55 games in 24 ms
55 games in 23 ms
55 games in 21 ms
55 games in 22 ms
55 games in 18 ms
55 games in 22 ms
55 games in 23 ms
Average = 420 microseconds per game
          2380 games per second
```
Much faster compare to the previous implementation:
```
55 games in 49 ms
55 games in 45 ms
55 games in 47 ms
55 games in 81 ms
55 games in 44 ms
55 games in 41 ms
55 games in 46 ms
55 games in 41 ms
55 games in 83 ms
55 games in 46 ms
Average = 950 microseconds per game
          1052 games per second
```
And old `scala-parser-combinators` implementation:
```
55 games in 127 ms
55 games in 101 ms
55 games in 89 ms
55 games in 82 ms
55 games in 80 ms
55 games in 66 ms
55 games in 63 ms
55 games in 61 ms
55 games in 64 ms
55 games in 92 ms
Average = 1500 microseconds per game
          666 games per second
```